reordered some Grid and Grispy tests
