+++ tests/test_brute_force.py
@@ -12,158 +12,7 @@ import numpy as np
 import numpy.testing as npt
 import pytest
 
-from grispy import Grid, GriSPy
-
-# =========================================================================
-# Test Grid class
-# =========================================================================
-
-
-class Test_Grid:
-    @pytest.fixture
-    def grid(self):
-        data = np.array(
-            [
-                [0, 0, 0],
-                [0, 0, 1],
-                [0, 1, 0],
-                [0, 1, 1],
-                [1, 0, 0],
-                [1, 0, 1],
-                [1, 1, 0],
-                [1, 1, 1],
-            ]
-        )
-        return Grid(data, N_cells=3)
-
-    @pytest.fixture
-    def valid_input(self):
-        rng = np.random.default_rng(1234)
-        d = dict()
-        # Define valid input data
-        d["in_points"] = rng.random((5, 3))
-        d["out_points"] = rng.random((5, 3)) + 10
-        d["mix_points"] = np.vstack((d["in_points"], d["out_points"]))
-        return d
-
-    def test_grid_property_edges(self, grid):
-        dmin = grid.data.min(axis=0) - grid.epsilon
-        dmax = grid.data.max(axis=0) + grid.epsilon
-        k_bins = np.linspace(dmin, dmax, grid.N_cells + 1)
-        expected = k_bins[[0, -1], :]
-        npt.assert_almost_equal(grid.edges, expected, 14)
-
-    @pytest.mark.parametrize(
-        "pname", ["in_points", "out_points", "mix_points"]
-    )
-    def test_grid_contains_inside(self, grid, valid_input, pname):
-        points = valid_input[pname]
-        lower = 0 < points
-        upper = points < 1
-        expected = (lower & upper).prod(axis=1, dtype=bool)
-        result = grid.contains(points)
-        npt.assert_equal(expected, result)
-
-    def test_cell_digits(self, grid):
-        expected = np.array(
-            [
-                [0, 0, 0],
-                [0, 0, 2],
-                [0, 2, 0],
-                [0, 2, 2],
-                [2, 0, 0],
-                [2, 0, 2],
-                [2, 2, 0],
-                [2, 2, 2],
-            ]
-        )
-        result = grid.cell_digits(grid.data)
-        npt.assert_equal(expected, result)
-
-    def test_cell_digits_outside(self, grid):
-        data = np.array(
-            [
-                [0, 0, 0],
-                [-1, 0, 1],
-                [0, 1, 0],
-                [0, 1, 1],
-                [1, 0, 0],
-                [1, 0, -1],
-                [1, 1, 0],
-                [1, 1, 1],
-            ]
-        )
-        expected = np.array(
-            [
-                [0, 0, 0],
-                [-1, -1, -1],
-                [0, 2, 0],
-                [0, 2, 2],
-                [2, 0, 0],
-                [-1, -1, -1],
-                [2, 2, 0],
-                [2, 2, 2],
-            ]
-        )
-        result = grid.cell_digits(data)
-        npt.assert_equal(expected, result)
-
-    def test_cell_id(self, grid):
-        expected = np.array([0, 18, 6, 24, 2, 20, 8, 26])
-        result = grid.cell_id(grid.data)
-        npt.assert_equal(expected, result)
-
-    def test_cell_id_outside(self, grid):
-        data = np.array(
-            [
-                [0, 0, 0],
-                [-1, 0, 1],
-                [0, 1, 0],
-                [0, 1, 1],
-                [1, 0, 0],
-                [1, 0, -1],
-                [1, 1, 0],
-                [1, 1, 1],
-            ]
-        )
-        expected = np.array([0, -1, 6, 24, 2, -1, 8, 26])
-        result = grid.cell_id(data)
-        npt.assert_equal(expected, result)
-
-    def test_cell_digits2id(self, grid):
-        digits = np.array(
-            [
-                [0, 0, 0],
-                [0, 0, 2],
-                [0, 2, 0],
-                [0, 2, 2],
-                [2, 0, 0],
-                [2, 0, 2],
-                [2, 2, 0],
-                [2, 2, 2],
-            ]
-        )
-        expected = np.array([0, 18, 6, 24, 2, 20, 8, 26])
-        result = grid.cell_digits2id(digits)
-        npt.assert_equal(expected, result)
-
-    def test_id2digits(self, grid):
-        ids = np.array([0, 18, 6, 24, 2, 20, 8, 26])
-        expected = np.array(
-            [
-                [0, 0, 0],
-                [0, 0, 2],
-                [0, 2, 0],
-                [0, 2, 2],
-                [2, 0, 0],
-                [2, 0, 2],
-                [2, 2, 0],
-                [2, 2, 2],
-            ]
-        )
-        result = grid.cell_id2digits(ids)
-        npt.assert_equal(expected, result)
-
+from grispy import GriSPy
 
 # =========================================================================
 # Test GriSPy class
